requirements.txt
streamlit
pandas
pdf2image
pillow
pytesseract
transformers
torch
accelerate
plotly
poppler-utils
torchvision
huggingface_hub